refactor reporting batch control routes
